% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixins.R
\name{array_to_pylist}
\alias{array_to_pylist}
\title{Translate from an R array to a python embedded list format.}
\usage{
array_to_pylist(ary, major = c("row", "column"))
}
\arguments{
\item{ary}{an at most 2d array}

\item{major}{should the array be column or row major?}
}
\value{
a string of the array in python list notation
}
\description{
Translate from an R array to a python embedded list format.
}
