% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-crunch-test.R
\name{cubify}
\alias{cubify}
\title{Turn readable test expectations into an array}
\usage{
cubify(..., dims)
}
\arguments{
\item{...}{any number of numerics to turn into an array}

\item{dims}{the dimensions of array to return}
}
\value{
an array with the contents of \code{...`` with the shape of}dims`
}
\description{
Turn readable test expectations into an array
}
