% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-redactor.R
\name{enable_crunch_redactor}
\alias{enable_crunch_redactor}
\title{A quick shortcut to enable httptest redaction for Crunch payloads}
\usage{
enable_crunch_redactor()
}
\description{
A quick shortcut to enable httptest redaction for Crunch payloads
}
