% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-crunch-test.R
\name{loadBareCube}
\alias{loadBareCube}
\title{Load a CrunchCube fixture from minimal file}
\usage{
loadBareCube(filename)
}
\arguments{
\item{filename}{a character filename to load the fixture from}
}
\value{
a CrunchCube
}
\description{
Sometimes tests were written with cube fixtures that lacked  the
\code{{"element": "shoji:view", "self": URL, "value":{}}} wrapper, this function
reads those in.
}
