% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-crunch-test.R
\name{loadCube}
\alias{loadCube}
\title{Load a CrunchCube fixture from file}
\usage{
loadCube(filename)
}
\arguments{
\item{filename}{a character filename to load the fixture from}
}
\value{
a CrunchCube
}
\description{
Load a CrunchCube fixture from file
}
