% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-redactor.R
\name{redact_crunch}
\alias{redact_crunch}
\title{A default redactor for Crunch payloads}
\usage{
redact_crunch(response)
}
\arguments{
\item{response}{the response to redact}
}
\description{
A default redactor for Crunch payloads
}
