% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload-test-data.R
\name{reload_test_data}
\alias{reload_test_data}
\title{Reload test data}
\usage{
reload_test_data()
}
\description{
Reloads all types, mrdf, mrdf with a multiple resonse and apidocs test data
to whichever backend is set in options.
}
