% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchdev.R
\name{setupCrunch}
\alias{setupCrunch}
\title{Setup Crunch to use a particular backend}
\usage{
setupCrunch(host, ...)
}
\arguments{
\item{host}{host to connect to}

\item{...}{passed to \code{\link[crunch:setCrunchAPI]{crunch::setCrunchAPI()}}}
}
\value{
nothing
}
\description{
If the host provided is in the \code{crunchdev.test.hosts} group, test authentication will be set (\code{crunchdev.test.auth}). Otherwise, user authentication will be set. If \code{host} is not recognized, it will be passed to \code{\link[crunch:setCrunchAPI]{crunch::setCrunchAPI()}}
}
