% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchdev.R
\name{shine_covr}
\alias{shine_covr}
\title{Check package coverage and shine the result}
\usage{
shine_covr(...)
}
\arguments{
\item{...}{options passed to \code{package_coverage}}
}
\value{
nothing
}
\description{
A convenience method for \code{covr::shine(covr::package_coverage())}
}
