% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchdev.R
\name{styleFile}
\alias{styleFile}
\title{Style file}
\usage{
styleFile(path)
}
\arguments{
\item{path}{Optional, the path of the file to style. If ommitted it is the
active RStudio document.}
}
\description{
Styles a file according to the crunch style guide.
}
