% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchdev.R
\name{test_crunch}
\alias{test_crunch}
\title{Function to configure and kick off tests of Crunch packages}
\usage{
test_crunch(filter = NULL, ...)
}
\arguments{
\item{filter}{a character of the regex of the filter to pre-populate
the test gadget input with}

\item{...}{options to pass to the test_gadget}
}
\value{
nothing
}
\description{
Gets document context, and then presents the test_gadget which has
configuration options for testing Crunch packages
}
